+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * `SBBSelect`: Adjusted paddings to match the current specifications
 * `SBBMultiSelect`: Adjusted paddings to match the current specifications
 * `SBBRadioButtonListItem`: Adjusted paddings to match the current specifications
+* `SBBSelect`: Field `label` is now optional because there is now a variant without label
 ### Deprecated
 ### Removed
 * `SBBListHeader`: Removed fields `icon` and `onCallToAction` to match the current specifications